Move pdf parsing logic to lib
